Fixed bug [ 1769832 ] No HTML file generated for sql_tables
The problem occurred for _all_ nested types, not just for 
sql_types.

Nested types are difficult to handle in full generality. 
The current solution requires that nested type names are 
unique within a package. The parser will complain if two 
compound types each define a nested type with the same name.